refactor: move the community route handlers into api/lib
api/community.ts and api/community/[id].ts were 726 and 754 code lines. The
routes now dispatch on method and delegate: publish and list handlers move to
api/lib/communityPublish and communityList; the design actions, mutations and
their shared guards move to api/lib/communityDesignActions, Mutations and
Shared. readOptionalSession, identical in three routes, moves to
api/lib/session, and the route tests' session mocks gain it.
