+++ api/lib/communityPublish.ts
@@ -1,0 +1,453 @@
+/** POST /api/community: publish a design, resolving lineage and refusing exact duplicates. */
+import { randomUUID } from 'node:crypto';
+import type { VercelRequest, VercelResponse } from '@vercel/node';
+import { del, put } from '@vercel/blob';
+import {
+  cardFromRecord,
+  communityDesignUrl,
+  DUPLICATE_DESIGN_RESPONSE,
+  REMIX_UNCHANGED_RESPONSE,
+} from './communityRecord.js';
+import { requireSession } from './session.js';
+import { checkRateLimit, getRedis } from './rateLimit.js';
+import { logger } from './logger.js';
+import { ErrorCode, rateLimited, sendError, serviceUnavailable } from './shared.js';
+import { parseCommunityLineage, validateCommunityPublish } from './communityValidation.js';
+import { deriveAuthorPublicId, generateCommunityDesignId } from './communityIds.js';
+import { checkCommunityPublishQuota } from './communityQuota.js';
+import {
+  adjustRemixCredit,
+  communityContentHash,
+  communityMeshBlobPath,
+  communityDesignContent,
+  communityParamsFingerprint,
+  communityThumbBlobPath,
+  deleteCommunityDesignBlob,
+  deriveAssemblyMetrics,
+  deriveCommunityMetrics,
+  isModeratedContent,
+  readCommunityCards,
+  readCommunityDesignBlob,
+  removeFromCommunityIndexes,
+  upsertCommunityIndexes,
+  writeCommunityCard,
+  writeCommunityDesignBlob,
+} from './communityStore.js';
+import { COMMUNITY_EXAMPLE_PARAM_HASHES } from './communityExampleParamHashes.js';
+import type { CommunityDesignRecord, CommunityLineage } from './communityStore.js';
+import {
+  communityAuthorKey,
+  communityChildrenKey,
+  communityDenylistKey,
+  communityDesignKey,
+  communityParamsHashKey,
+  communityPublishedKey,
+  communityPublishLockKey,
+} from './redisKeys.js';
+
+export type RedisClient = NonNullable<ReturnType<typeof getRedis>>;
+
+const FIRST_REVISION = 1;
+
+// Short per-user publish lock TTL: long enough to cover the check->write
+// sequence of one publish, short enough that a crashed request self-heals.
+const PUBLISH_LOCK_TTL_MS = 10_000;
+
+type LineageResolveResult =
+  | { ok: true; lineage: CommunityLineage | null; parentContent: Record<string, unknown> | null }
+  | { ok: false; message: string };
+
+/**
+ * Lineage display fields and the root itself come from the stored parent
+ * record, never the client: accepting client values would let a publisher
+ * fabricate "remix of X by Y" or "originally by Z" credit for arbitrary real
+ * authors. The parent must be live at publish time and its stored chain
+ * dictates the root; a client rootId that disagrees is rejected. When the
+ * chain's root is legitimately deleted, its snapshot survives via the parent
+ * record.
+ *
+ * ACCEPTED RISK (A11, community-showcase-plan): the parent link carries no
+ * proof-of-remix. Any live design can be claimed as a parent; only the credit
+ * text is server-derived, so a spammer could inflate a victim's remix count.
+ * This is bounded by the 10/day publish rate limit and the report path, and
+ * further shrunk by the remix-must-differ rule (B4) and the exact-duplicate
+ * guard (B3). We deliberately do not build proof-of-remix.
+ */
+async function resolveLineage(
+  redis: RedisClient,
+  lineage: CommunityLineage | null
+): Promise<LineageResolveResult> {
+  if (lineage === null) return { ok: true, lineage: null, parentContent: null };
+  const parentCard = (await readCommunityCards(redis, [lineage.parentId]))[0] ?? null;
+  if (parentCard === null || parentCard.status !== 'live') {
+    return { ok: false, message: 'lineage.parentId does not reference a live design' };
+  }
+  const parentRecord = await readCommunityDesignBlob(lineage.parentId);
+  if (parentRecord === null) {
+    return { ok: false, message: 'lineage.parentId does not reference a live design' };
+  }
+  const rootId = parentRecord.lineage?.rootId ?? lineage.parentId;
+  if (lineage.rootId !== rootId) {
+    return { ok: false, message: 'lineage.rootId does not match the parent design chain' };
+  }
+  let rootAuthorName = parentRecord.lineage?.rootAuthorName ?? parentCard.authorName;
+  if (rootId !== lineage.parentId) {
+    const rootCard = (await readCommunityCards(redis, [rootId]))[0] ?? null;
+    if (rootCard !== null && rootCard.status === 'live') {
+      rootAuthorName = rootCard.authorName;
+    }
+  }
+  return {
+    ok: true,
+    lineage: {
+      parentId: lineage.parentId,
+      rootId,
+      parentName: parentCard.name,
+      parentAuthorName: parentCard.authorName,
+      rootAuthorName,
+    },
+    parentContent: communityDesignContent(parentRecord),
+  };
+}
+
+async function findPublishedIdByContentHash(
+  redis: RedisClient,
+  userId: string,
+  contentHash: string
+): Promise<string | null> {
+  const publishedIds = await redis.smembers(communityPublishedKey(userId));
+  if (publishedIds.length === 0) return null;
+  const pipeline = redis.pipeline();
+  for (const id of publishedIds) {
+    pipeline.hget(communityDesignKey(id), 'contentHash');
+  }
+  // A swallowed exec failure would disable idempotency and mint duplicate
+  // designs while consuming quota, so an unhealthy redis must fail the
+  // publish instead.
+  const results = await pipeline.exec();
+  if (results === null) {
+    throw new Error('Community idempotency check failed: redis connection lost');
+  }
+  for (let i = 0; i < publishedIds.length && i < results.length; i++) {
+    const [error, value] = results[i];
+    if (error) {
+      throw new Error(`Community idempotency check failed: ${error.message}`);
+    }
+    if (value === contentHash) {
+      // A8: never resurface a hidden/removed design as a 200 idempotency hit.
+      // A moderated design keeping its content hash would otherwise let a
+      // re-publish return the id of a design the gallery no longer shows.
+      const status = await redis.hget(communityDesignKey(publishedIds[i]), 'status');
+      if (status === 'live') return publishedIds[i];
+    }
+  }
+  return null;
+}
+
+/**
+ * B3 exact-duplicate guard. True when the sanitized params re-upload a built-in
+ * example, or match a currently-live design published by a different author.
+ * The author's own re-publish never reaches here (idempotency returns first).
+ */
+async function isExactDuplicate(
+  redis: RedisClient,
+  contentFingerprint: string,
+  authorPublicId: string
+): Promise<boolean> {
+  if (COMMUNITY_EXAMPLE_PARAM_HASHES.has(contentFingerprint)) return true;
+  const candidateId = await redis.get(communityParamsHashKey(contentFingerprint));
+  if (candidateId === null || candidateId === '') return false;
+  const [status, candidateAuthor] = await redis.hmget(
+    communityDesignKey(candidateId),
+    'status',
+    'authorPublicId'
+  );
+  return status === 'live' && candidateAuthor !== null && candidateAuthor !== authorPublicId;
+}
+
+export async function handlePublish(req: VercelRequest, res: VercelResponse): Promise<void> {
+  if (process.env.COMMUNITY_PUBLISH_ENABLED !== 'true') {
+    serviceUnavailable(res, 'Community publishing is not available.');
+    return;
+  }
+
+  const session = await requireSession(req, res);
+  if (!session) return;
+
+  try {
+    const rate = await checkRateLimit(session.userId, 'community.publish');
+    if (!rate.allowed) {
+      rateLimited(res, rate.retryAfterSeconds, 'Publish limit reached. Try again later.');
+      return;
+    }
+
+    const redis = getRedis();
+    if (!redis) {
+      serviceUnavailable(res);
+      return;
+    }
+
+    const authorPublicId = deriveAuthorPublicId(session.userId);
+    if (authorPublicId === null) {
+      logger.error('Community publish failed: TOKEN_SALT is not configured');
+      serviceUnavailable(res);
+      return;
+    }
+
+    const validated = validateCommunityPublish(req.body);
+    if (!validated.valid) {
+      res.status(400).json({ error: validated.error.message, code: validated.error.code });
+      return;
+    }
+    const payload = validated.payload;
+
+    const lineageParse = parseCommunityLineage((req.body as Record<string, unknown>).lineage);
+    if (!lineageParse.ok) {
+      res.status(400).json({ error: lineageParse.message, code: 'INVALID_LINEAGE' });
+      return;
+    }
+
+    const denied = await redis.sismember(communityDenylistKey(), session.userId);
+    if (denied === 1) {
+      // Deliberately neutral: the response must not reveal deny-listing.
+      sendError(res, 403, ErrorCode.UNAUTHORIZED, 'Publishing is not available for this account.');
+      return;
+    }
+
+    // A9: serialize one user's overlapping publishes so the check->write
+    // sequence cannot interleave into a duplicate mint or a quota overshoot.
+    // The value is a unique token: if this publish outlives PUBLISH_LOCK_TTL_MS
+    // and a second acquires the lock, our release must not delete the second
+    // request's lock (fencing), so release is a compare-and-delete on the token.
+    const lockKey = communityPublishLockKey(session.userId);
+    const lockToken = randomUUID();
+    const lockAcquired = await redis.set(lockKey, lockToken, 'PX', PUBLISH_LOCK_TTL_MS, 'NX');
+    if (lockAcquired !== 'OK') {
+      res.status(409).json({
+        error: 'Another publish is already in progress. Try again in a moment.',
+        code: 'PUBLISH_IN_PROGRESS',
+      });
+      return;
+    }
+
+    try {
+      const contentFingerprint = communityParamsFingerprint(communityDesignContent(payload));
+      const contentHash = communityContentHash({
+        content: communityDesignContent(payload),
+        name: payload.name,
+        description: payload.description,
+        category: payload.category,
+        authorName: payload.authorName,
+        lineage: lineageParse.lineage,
+      });
+      // Content taken down by moderation cannot come back, and that outranks
+      // every check below including idempotency. The dedupe checks deliberately
+      // only match LIVE designs, so a hidden original is invisible to them and
+      // a re-post of the identical payload would otherwise mint a fresh design
+      // with an empty reports set. The tombstone is keyed by content, so it
+      // survives the design — and the whole account — being deleted. Neutral
+      // 403, matching the deny-list response: the reply must not confirm that a
+      // takedown happened.
+      if (await isModeratedContent(redis, contentHash)) {
+        sendError(res, 403, ErrorCode.UNAUTHORIZED, 'Publishing is not available for this design.');
+        return;
+      }
+
+      // Idempotency runs before quota: a retry of an already-published design
+      // must return its id even when the user sits at the live-design cap.
+      const existingId = await findPublishedIdByContentHash(redis, session.userId, contentHash);
+      if (existingId !== null) {
+        res.status(200).json({ id: existingId, url: communityDesignUrl(existingId) });
+        return;
+      }
+
+      // B3: reject a verbatim re-upload of a built-in example or of another
+      // author's live design. The author's own re-publish returned above.
+      if (await isExactDuplicate(redis, contentFingerprint, authorPublicId)) {
+        res.status(409).json(DUPLICATE_DESIGN_RESPONSE);
+        return;
+      }
+
+      const quota = await checkCommunityPublishQuota(redis, session.userId);
+      if (!quota.ok) {
+        sendError(
+          res,
+          413,
+          ErrorCode.SIZE_LIMIT,
+          `Published design limit reached (${quota.error.limit} live designs).`
+        );
+        return;
+      }
+
+      const lineageResolve = await resolveLineage(redis, lineageParse.lineage);
+      if (!lineageResolve.ok) {
+        res.status(400).json({ error: lineageResolve.message, code: 'INVALID_LINEAGE' });
+        return;
+      }
+      const lineage = lineageResolve.lineage;
+
+      // B4: a remix must differ from its parent, or remix credit could be
+      // farmed by re-uploading the parent unchanged.
+      if (
+        lineage !== null &&
+        lineageResolve.parentContent !== null &&
+        communityParamsFingerprint(lineageResolve.parentContent) === contentFingerprint
+      ) {
+        res.status(409).json(REMIX_UNCHANGED_RESPONSE);
+        return;
+      }
+
+      const id = generateCommunityDesignId();
+      const now = Date.now();
+      const [thumbBlobs, meshBlob] = await Promise.all([
+        Promise.all(
+          payload.thumbnails.map((thumbnail, angle) =>
+            put(
+              communityThumbBlobPath(id, FIRST_REVISION, angle),
+              Buffer.from(thumbnail, 'base64'),
+              {
+                access: 'public',
+                contentType: 'image/webp',
+                addRandomSuffix: false,
+                allowOverwrite: false,
+              }
+            )
+          )
+        ),
+        put(communityMeshBlobPath(id, FIRST_REVISION), Buffer.from(payload.glb, 'base64'), {
+          access: 'public',
+          contentType: 'model/gltf-binary',
+          addRandomSuffix: false,
+          allowOverwrite: false,
+        }),
+      ]);
+
+      const record: CommunityDesignRecord = {
+        id,
+        authorPublicId,
+        authorName: payload.authorName,
+        name: payload.name,
+        description: payload.description,
+        category: payload.category,
+        techniques: payload.techniques,
+        ...(payload.kind === 'assembly'
+          ? { kind: 'assembly' as const, envelope: payload.envelope, structure: payload.structure }
+          : { params: payload.params }),
+        metrics:
+          payload.kind === 'assembly'
+            ? deriveAssemblyMetrics(payload.envelope ?? {}, payload.heightUnits ?? 1)
+            : deriveCommunityMetrics(payload.params ?? {}),
+        lineage,
+        thumbnails: thumbBlobs.map((blob) => blob.url),
+        meshUrl: meshBlob.url,
+        photos: [],
+        featured: false,
+        createdAt: now,
+        updatedAt: now,
+        status: 'live',
+      };
+
+      try {
+        await writeCommunityDesignBlob(record);
+        await writeCommunityCard(redis, cardFromRecord(record));
+        await redis.hset(communityDesignKey(id), { contentHash });
+        await upsertCommunityIndexes(redis, id, {
+          createdAt: now,
+          remixes: 0,
+          likes: 0,
+          prints: 0,
+        });
+        await redis.sadd(communityPublishedKey(session.userId), id);
+        await redis.sadd(communityAuthorKey(authorPublicId), id);
+        // Recorded at publish so remix claims are auditable and so the delete
+        // path's srem from the parent's children set has something to undo.
+        if (lineage !== null) {
+          await redis.sadd(communityChildrenKey(lineage.parentId), id);
+        }
+      } catch (writeErr) {
+        // The thumbnails and GLB were already uploaded to public, predictable
+        // paths; without this cleanup a failed publish would strand CDN assets
+        // that no record references and no purge can enumerate.
+        const uploadedAssets = [...thumbBlobs.map((blob) => blob.url), meshBlob.url];
+        await del(uploadedAssets).catch((delErr: unknown) => {
+          logger.error('Rollback failed: orphan community assets left after publish failure', {
+            id,
+            error: delErr instanceof Error ? delErr.message : String(delErr),
+          });
+        });
+        await deleteCommunityDesignBlob(id).catch((delErr: unknown) => {
+          logger.error('Rollback failed: orphan community design blob left after publish failure', {
+            id,
+            error: delErr instanceof Error ? delErr.message : String(delErr),
+          });
+        });
+        // Redis writes may have partially landed before the failure; without
+        // this cleanup a stranded card hash surfaces in the gallery list while
+        // the detail 404s, and the published-set entry burns a quota slot.
+        await (async () => {
+          const cleanup = redis.pipeline();
+          cleanup.del(communityDesignKey(id));
+          cleanup.srem(communityPublishedKey(session.userId), id);
+          cleanup.srem(communityAuthorKey(authorPublicId), id);
+          if (lineage !== null) {
+            cleanup.srem(communityChildrenKey(lineage.parentId), id);
+          }
+          await cleanup.exec();
+          await removeFromCommunityIndexes(redis, id);
+        })().catch((cleanupErr: unknown) => {
+          logger.error(
+            'Rollback failed: orphan community redis entries left after publish failure',
+            {
+              id,
+              error: cleanupErr instanceof Error ? cleanupErr.message : String(cleanupErr),
+            }
+          );
+        });
+        throw writeErr;
+      }
+
+      // Post-commit bookkeeping: the design is already live, so these keep the
+      // duplicate index and remix stats in sync best-effort. A failure here
+      // must not fail an already-successful publish.
+      await redis.set(communityParamsHashKey(contentFingerprint), id).catch((indexErr: unknown) => {
+        logger.warn('Community publish: params-hash index write failed', {
+          id,
+          error: indexErr instanceof Error ? indexErr.message : String(indexErr),
+        });
+      });
+      // A1: crediting the parent and root remix counts (and rescoring the
+      // most-remixed index) is the counterpart of the delete-path decrement.
+      if (lineage !== null) {
+        await adjustRemixCredit(redis, lineage.parentId, lineage.rootId, 1).catch(
+          (remixErr: unknown) => {
+            logger.warn('Community publish: remix credit bump failed', {
+              id,
+              error: remixErr instanceof Error ? remixErr.message : String(remixErr),
+            });
+          }
+        );
+      }
+
+      res.status(201).json({ id, url: communityDesignUrl(id) });
+    } finally {
+      // Compare-and-delete: only release the lock if we still hold it, so a
+      // publish that overran the TTL cannot delete a newer publish's lock.
+      await redis
+        .eval(
+          "if redis.call('get', KEYS[1]) == ARGV[1] then return redis.call('del', KEYS[1]) else return 0 end",
+          1,
+          lockKey,
+          lockToken
+        )
+        .catch(() => {
+          // The lock self-heals via its PX TTL; a failed release is not fatal.
+        });
+    }
+  } catch (error) {
+    logger.error('Community publish error', {
+      error: error instanceof Error ? error.message : String(error),
+      stack: error instanceof Error ? error.stack : undefined,
+    });
+    sendError(res, 500, ErrorCode.SERVER_ERROR, 'Failed to publish design');
+  }
+}
